+++ src/components/AboutAppAccordian/AboutAppAccordianStyle.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const AccordionContainer = styled.div`
+  width: 100%;
+  padding: 32px 8px 8px 8px;
+  box-sizing: border-box;
+`;
