+++ src/components/ArticleCard/ArticleCardStyle.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const CardContainer = styled.div`
+  margin-top: 16px;
+  border-bottom: 1px solid lightgray;
+  height: 112px;
+  cursor: pointer;
+`;
+
+export const TextHighlight = styled.span`
+  color: grey;
+  font-weight: 400;
+`;
