+++ src/components/Header/HeaderStyle.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const HeaderContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  padding: 8px 16px;
+  box-sizing: border-box;
+  box-shadow: 0 4px 2px -2px lightgray;
+  width: 100%;
+`;
