+++ src/views/Article/Article.jsx
@@ -1,16 +1,10 @@
-import {
-  Button,
-  Card,
-  CardActions,
-  CardContent,
-  Typography,
-} from "@mui/material";
+import { Button, CardActions, CardContent, Typography } from "@mui/material";
 import React from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import styled from "styled-components";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
 import { useQuery } from "@tanstack/react-query";
 import { CenteredLoading } from "../../styles/styled";
+import { ArticleCard } from "./ArticleStyle";
 
 const fetchArticle = async (id) => {
   const res = await fetch(`http://127.0.0.1:5000/api/blogpost/${id}`, {
@@ -68,8 +62,3 @@ const Article = () => {
 };
 
 export default Article;
-
-const ArticleCard = styled(Card)`
-  margin: 32px 8px 0 8px;
-  box-sizing: border-box;
-`;
