+++ src/views/Article/ArticleStyle.js
@@ -1,0 +1,7 @@
+import { Card } from "@mui/material";
+import styled from "styled-components";
+
+export const ArticleCard = styled(Card)`
+  margin: 32px 8px 0 8px;
+  box-sizing: border-box;
+`;
