+++ src/views/CreatePosts/CreatePostsStyle.js
@@ -1,0 +1,11 @@
+import { TextField } from "@mui/material";
+import styled from "styled-components";
+
+export const MarginedTextField = styled(TextField)`
+  margin-bottom: 32px;
+`;
+
+export const FormContainer = styled.div`
+  margin-top: 32px;
+  height: 100%;
+`;
